+++ tests/Unit/DSC_NetIPInterface.Tests.ps1
@@ -163,6 +163,14 @@ try
                 ParameterFilter = {
                     $InterfaceAlias -eq 'Ethernet' -and $AddressFamily -eq 'IPv4' -and $WeakHostSend -eq 'Disabled'
                 }
+            },
+            @{
+                Name            = 'NlMtu'
+                MockedValue     = [System.Uint32] 1600
+                TestValue       = [System.Uint32] 1500
+                ParameterFilter = {
+                    $InterfaceAlias -eq 'Ethernet' -and $AddressFamily -eq 'IPv4' -and $NlMtuBytes -eq 1500
+                }
             }
         )
 
@@ -239,7 +247,7 @@ try
 
                     $script:result = Test-TargetResource @script:netIPInterfaceExists @comparisonParameter
 
-                    $script:result | Should -Be $false
+                    $script:result | Should -BeFalse
                 }
             }
 
@@ -261,7 +269,7 @@ try
 
                     $script:result = Test-TargetResource @script:netIPInterfaceExists @comparisonParameter
 
-                    $script:result | Should -Be $true
+                    $script:result | Should -BeTrue
                 }
             }
         }
@@ -327,7 +335,7 @@ try
 
                 It 'Should throw expected exception' {
                     $errorRecord = Get-InvalidOperationRecord `
-                        -Message ($script:localizedData.NetIPInterfaceDoesNotExistMessage -f $script:netIPInterfaceMissing.InterfaceAlias, $script:netIPInterfaceMissing.AddressFamily)
+                        -Message ($script:localizedData.NetworkIPInterfaceDoesNotExistMessage -f $script:netIPInterfaceMissing.InterfaceAlias, $script:netIPInterfaceMissing.AddressFamily)
 
                     {
                         $script:result = Get-NetworkIPInterface @script:netIPInterfaceMissing